changed Main.java and BmiService.java

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(double weightKg, double heightM) {
-        return (int) (weightKg / (heightM * heightM));  // метод возвращает целое число
+    public int calculate(double weightKg, double heightM) {
+        return (int) (weightKg / (heightM * heightM));
     }
 }
